twitter/requirements.txt
apify-client>=1.0.0
requests>=2.28.0
python-dotenv>=0.20.0 
streamlit>=1.22.0
pandas>=1.5.3
pillow>=9.5.0 
jinja2>=3.0.0